extract clean_phone
