+++ starfish/members/forms.py
@@ -1,10 +1,9 @@
-from string import digits
-
 from django import forms
+from members.models import PendingMember, get_by_email
 from partners.models import PartnerCampaign
 from regions.models import Zip
 
-from .models import PendingMember, get_by_email
+from starfish.helpers.nanp_phone import clean_nanp_phone
 
 PHONE_PUNCTUATION = r".()- "
 PHONE_TRANSLATION_TABLE = str.maketrans("", "", PHONE_PUNCTUATION)
@@ -73,113 +72,5 @@ class PendingMemberForm(forms.ModelForm):
         return email
 
     def clean_phone(self):
-        """Validate NANP Telephone Numbers
-
-        https://en.wikipedia.org/wiki/North_American_Numbering_Plan#Modern_plan
-
-        Telephone numbers have the following form: NXX-NXX-XXXX per the "North
-        American Numbering Plan"
-
-        - Exactly 10 digits
-        - N must be 2-9
-        - X may be 0-9 with some exceptions.
-
-        Additional rules:
-
-        - First NXX is the area code.
-            - Reserved and non-geographic codes.
-                - NXX where the two X are identical, reserved for non-geographic
-                  use. 911, 988, 800, etc.
-                - N90-N99 are reserved for expansion.
-                - 370-379, 960-969 are reserved for expansion.
-                - 880-889 are reserved for toll-free use.
-                - 950 is reserved for usage of prepaid calling cards.
-                - 958, 959 are reserved for system testing.
-        - Second NXX is the prefix (or exchange).
-            - Prefixes of the form N11 were reserved for "Easily Recognizable
-              Codes" during the transition to the modern plan and, while we must
-              dial with 10 digits now, this rule has stuck around.
-            - Prefix 555 has a block of line numbers reserved for fictional use.
-              Only two line numbers (1212, 4334) are currently in use, the
-              others are reserved until a need is approved by the Industry
-              Numbering Committee:
-              https://www.nanpa.com/numbering/555-line-numbers. We can reject
-              any of these numbers at this time.
-        - XXXX is the line number. There are no restrictions.
-        """
         phone_input = self.cleaned_data.get("phone")
-
-        if not phone_input:
-            return None
-
-        if not isinstance(phone_input, str):
-            msg = "Phone number not a string."
-            raise RuntimeError(msg)
-
-        VERIFICATION_MSG = (
-            "Please enter a 10-digit North American phone number without the 1."
-            "\n Spaces and the following punctuation are allowed: ().-"
-        )
-
-        if not _is_phone_input_only_allowed_characters(phone_input):
-            raise forms.ValidationError(VERIFICATION_MSG)
-
-        phone_input = _remove_punctuation_and_space(phone_input)
-
-        if not len(phone_input) == 10:
-            raise forms.ValidationError(VERIFICATION_MSG)
-
-        area_code = phone_input[:3]
-        if not _is_phone_area_code_valid(area_code):
-            msg = (
-                "Please ensure your phone number has a valid area code."
-                "\nArea codes cannot start with 0 or 1."
-            )
-            raise forms.ValidationError(msg)
-
-        prefix = phone_input[3:6]
-        if not _is_phone_prefix_valid(prefix):
-            msg = (
-                "Please ensure your three-digit prefix (exchange) is valid."
-                "\nPrefixes cannot start with 0 or 1, and cannot be 555."
-            )
-            raise forms.ValidationError(msg)
-
-        return phone_input
-
-
-def _is_phone_input_only_allowed_characters(v):
-    return set(v) <= set(PHONE_PUNCTUATION + digits)
-
-
-def _remove_punctuation_and_space(v):
-    return v.translate(PHONE_TRANSLATION_TABLE)
-
-
-def _is_phone_area_code_valid(v):
-    return _is_phone_nanp_n_digit_valid(v) and _is_phone_area_code_geographic(v)
-
-
-def _is_phone_area_code_geographic(v):
-    return not (
-        _is_phone_easily_recognizable_code(v)
-        or v[1] == "9"  # Reserved for 12-digit expansion plan
-        or v[:2] in ("37", "88", "96")  # Reserved for non-geographic use
-        or v in ("950", "958", "959")  # Reserved for non-geographic use
-    )
-
-
-def _is_phone_prefix_valid(v):
-    return _is_phone_nanp_n_digit_valid(v) and not _is_phone_prefix_reserved(v)
-
-
-def _is_phone_prefix_reserved(v):
-    return v == "555"  # Reserved for special use
-
-
-def _is_phone_easily_recognizable_code(v):
-    return v[1] == v[2]
-
-
-def _is_phone_nanp_n_digit_valid(v):
-    return v[0] not in ("0", "1")
+        return clean_nanp_phone(phone_input, forms.ValidationError)
